folded Node class into HasTag
We do something a little wonky now. HasTag attaches itself to the
underlying XML tag through 'tag.__owner = self'.

This is a huge change. Elements no longer need to keep track of
their own parents/children, because that information can now
be determined from the HTML tree (figuring that any HTML tag with
a `__owner` attribute belongs to an Element).