remove global variable from server
Previously, the GUI currently being served was stored as the
global variable CURRENT_GUI. Therefore, only one GUI could be
run at a time.

Now, when a GUI is served, it's given its own RequestHandler class,
which stores the GUI as a class-level variable, removing the need
for global variables and allowing multiple GUIs to be served at once
(on different ports).